refactor(Worker): No need for separate folder yet
